ConfirmActionButton component
The confirm action button component let specify the style of trigger button and inner confirmation message.
It accepts a message to show to the user.
It accepts a custom label for the confirmation button.

The component shows a confirmation modal to the user.

The component accepts a callback action that is triggered upon user click on confirmation button within modal.